refactor(speech): extract device recovery + model lifecycle from ParakeetEngine
Splits ParakeetEngine.swift (3679 lines, drifted from the spec's assumed 3598)
into three files, pure code motion with no behavior change:

- ParakeetDeviceRecovery.swift: device-change detection (CoreAudio default-input
  listener, AVAudioEngineConfigurationChange observer) and the
  attemptDeviceRecovery/scheduleConfigRecoveryTimeout executor. The pure decision
  tables this executor consults (ParakeetDeviceRecoveryReadinessPolicy,
  ParakeetDeviceRecoveryFailurePolicy, ParakeetDeviceRecoveryTimeoutPolicy)
  already existed as tested, Foundation-pure types in
  ParakeetStartRecordingFailurePolicy.swift — this file is the side-effecting
  executor around them, as spec W2-C's design intent asked for.
- ParakeetModelLifecycle.swift: model load/download/warmup/teardown paths
  (initialize, performInitialize, prefetchModelFilesIfNeeded, initializeEouModel,
  migrateLegacyParakeetCacheIfNeeded, bundledModelPath, plus new
  cancelModelWork()/teardownModel() helpers factored out of cleanup()).
- ParakeetEngine.swift drops from 3679 to 2793 lines and keeps its public API
  as owner; the two new files are internal collaborator extensions.

Adds the first dedicated unit-test suite for the device-recovery decision table
(Tests/DeviceRecoveryPolicyTests.swift, 9 cases across all three policy types),
registered in Tests/FastTests.manifest.

Several source-text drift-guard/contract tests (BluetoothRouteContractTests,
RepoCommandContractTests, DictationAudioRecoveryTests,
ParakeetStartRecordingFailurePolicyTests) assert on exact "private func X"
strings and file locations; updated them to match the moved functions' new
file and the access-level changes required to call across the new file
boundary (private -> internal on methods/properties whose only external
callers moved to a different file).

Deviation from spec: file lands at 2793 lines, short of the <2200 target.
The two extracted concerns (device recovery ~500 lines, model lifecycle
~360 lines) are fully and faithfully extracted per the design; the remaining
~2800 lines are recording/transcription core (installTapAndStartEngine,
startAudioWatchdog, stopRecording, transcribe, transcribeSamples) that the
spec's design explicitly keeps on ParakeetEngine as owner, not part of this
wave's scope.

Local build/fast-test verification was queued out by heavy machine
contention (13+ concurrent flock waiters on this box); CI is the gate for
this PR.

Cites codebase audit 2026-07-08 wave 2, spec W2-C.
