+++ Tests/FastTests.manifest
@@ -130,6 +130,7 @@ ParakeetPrewarmPolicyTests.swift:testParakeetPrewarmPolicy
 ParakeetRecoveryStateTests.swift:testParakeetRecoveryState
 ParakeetModelInitDiagnosticsTests.swift:testParakeetModelInitDiagnostics
 ParakeetStartRecordingFailurePolicyTests.swift:testParakeetStartRecordingFailurePolicy
+DeviceRecoveryPolicyTests.swift:testDeviceRecoveryPolicy
 MeetingRecordingStartGateTests.swift:testMeetingRecordingStartGate
 QuitConfirmationPreferencesTests.swift:testQuitConfirmationPreferences
 AutoCallDetectionPreferencesTests.swift:testAutoCallDetectionPreferences
